Add auth module and service
Added auth module, which implements UsersModule.
Added auth service, which contains a validateUser method, that accepts a
given username and password and tries to find the user based on the
username. If the user was found in the database, the method will compare
user's password with the provided password and return the id and name of
the user upon validation. If the username was not found or the password
does not match the provided one, the method returns null.